+++ src/fileIO/Snipper.cpp
@@ -1,42 +1,33 @@
 // This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-#include <mutex>
-#include <map>
-
 #include "fileIO/Snipper.hpp"
 
 namespace fs = std::filesystem;
-using file = std::pair<std::string, std::filesystem::file_time_type>;
-
-struct tnt::Snipper::impl
-{
-    std::recursive_mutex mtx;
-    std::map<std::string, file, std::less<>> files;
-};
 
 void tnt::Snipper::watchFile(std::string_view name)
 {
-    std::lock_guard lock{data->mtx};
-    std::string key{name.data()};
-    if (fs::exists(key))
-        data->files.try_emplace(key, key, fs::last_write_time(key));
+    mtx.lock();
+    hashed_string key{name.data()};
+    if (fs::exists(key.data()))
+        files.try_emplace(key.hash(), key.data(), fs::last_write_time(key.data()));
+    mtx.unlock();
 }
 
 bool tnt::Snipper::isModified(std::string_view file) noexcept
 {
-    std::lock_guard lock{data->mtx};
-    std::string const key{file.data()};
-    if (data->files.find(key) == data->files.cend())
+    std::lock_guard lock{mtx};
+    hashed_string key{file.data()};
+    if (files.find(key.hash()) == files.cend())
     {
-        data->files.try_emplace(key, key, fs::last_write_time(key));
+        files.try_emplace(key.hash(), key.data(), fs::last_write_time(key.data()));
         return false; // since we just added the file, there's no chance it has been modified
     }
 
-    if (auto time{fs::last_write_time(key)};
-        time != data->files[key].second)
+    if (auto time{fs::last_write_time(key.data())};
+        time != files[key.hash()].second)
     {
-        data->files[key].second = time;
+        files[key.hash()].second = time;
         return true;
     }
     return false;
@@ -44,7 +35,7 @@ bool tnt::Snipper::isModified(std::string_view file) noexcept
 
 void tnt::Snipper::unwatchFile(std::string_view filename) noexcept
 {
-    std::lock_guard lock{data->mtx};
-    if (auto it{data->files.find(filename.data())}; it != data->files.cend())
-        it = data->files.erase(it);
+    std::lock_guard lock{mtx};
+    if (auto it{files.find(fnv1a<char>(filename.data()))}; it != files.cend())
+        it = files.erase(it);
 }